requirements.txt
apache-airflow[celery]==2.10.5
python-binance
SQLAlchemy
pandas
psycopg2